system deadlock fix, blocks every 15 secs, int balances, score b^m, balances sum fix

+++ src/main/scala/cpos/model/Account.scala
@@ -2,7 +2,7 @@ package cpos.model
 
 import cpos.model.TypesAndConstants._
 
-case class Account(balance: Long, publicKey: PublicKey) {
+case class Account(balance: Int, publicKey: PublicKey) {
   require(balance >= 0)
   require(publicKey.length == PubKeyLength)
 }
